requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
filelock==3.18.0
Flask==3.1.1
fsspec==2025.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
setuptools==80.9.0
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
typing_extensions==4.14.1
Werkzeug==3.1.3
